Revert previous commit and change time dim name in readwrite_obs.F90
Several other subroutines in testsuite/src/lorenz96 expect the dim
name to be 'timesteps'. Therefore the dim name change was reverted in
generate_obs.F90 and the plot scripts and the name was changed in
readwrite_obs.F90 from 'steps' to 'timesteps'. Also, an integer
variable 'steps' was added to the output in readwrite_obs.F90 in
accordance with generate_obs.F90.