Development using filter 8 for soil moisture data assimilation(localized EnKF)
in the current version, the localized covariance is not evaluated if CLM
SA is used, in the version attached I added the corresponding evaluation.
init_dim_obs_pdaf.F90 needs to be modified as well because we need to
call a function there to
initialize the arrays used in the calculation in
localize_covar_pdaf.f90.