`add_obs_error_pdaf`: change `dim_obs_p` to `dim_obs`
This does not change the functionality, simply renaming a local
variable.

Reason: `add_obs_error_pdaf` is called by EnKF/LEnKF with the global
number of observations at a particular observation time and this
number is called `dim_obs` in the `src` routines, while `dim_obs_p` is
used for the number of PE-local observations.